requirements.txt
tgcrypto==1.2.5
pyrogram==2.0.106
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
